refactor(storage): 스토리지 모듈 디렉토리 구조 변경
- `data` 디렉토리 하위의 `data_source`와 `repository_impl`을 `storage` 모듈의 최상위로 이동
